Initial commit of the Fileset component
This component is extracted from the phpDocumentor main
repository as it is highly reusable for any project that needs
to collect a set of files based on a set of given parameters. It
supports ignoring based on a mask, globs and filtering based
on file extension.

The component is not highly unit tested in this commit but all
groundwork is there to test this component